Updated Queries

+++ packages/twenty-front/src/modules/users/graphql/queries/getCampaignList.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const GET_CAMPAIGN_LISTS = gql`
-query FindManyCampaigns($filter: CampaignFilterInput, $orderBy: CampaignOrderByInput, $lastCursor: String, $limit: Float) {
+query FindManyCampaigns($filter: CampaignFilterInput, $orderBy: CampaignOrderByInput, $lastCursor: String, $limit: Int) {
   campaigns(filter: $filter, orderBy: $orderBy, first: $limit, after: $lastCursor) {
     edges {
       node {
