+++ packages/twenty-front/src/modules/users/graphql/queries/getSpecialtyDetails.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client';
 
 export const GET_SPECIALTY = gql`
-query FindManySubspecialties($filter: SubspecialtyFilterInput, $orderBy: SubspecialtyOrderByInput, $lastCursor: String, $limit: Float) {
+query FindManySubspecialties($filter: SubspecialtyFilterInput, $orderBy: SubspecialtyOrderByInput, $lastCursor: String, $limit: Int) {
   subspecialties(
     filter: $filter
     orderBy: $orderBy
